// chpqa-app/src/app/api-services/chpqa-api/generated/model/replySubmission.ts
/**
 * WebApi
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Policy } from './policy';
import { RequestShceme } from './requestShceme';
import { ReplySubmissionGroups } from './replySubmissionGroups';


export interface ReplySubmission { 
    id?: string | null;
    name?: string | null;
    submissionFormType?: number | null;
    year?: string | null;
    version?: string | null;
    scheme?: RequestShceme;
    policyChpqa?: Policy;
    policyRocsCfd?: Policy;
    status?: number | null;
    schemeApplyingForCertificationIsExisting?: boolean | null;
    chpFuelBillingPeriod?: number | null;
    chpFuelBillingPeriodOther?: string | null;
    schemeLineDiagramExists?: boolean | null;
    schemeEnergyFlowDiagramExists?: boolean | null;
    annualHeatProfileExists?: boolean | null;
    dailyHeatProfileExists?: boolean | null;
    heatLoadDurationCurveExists?: boolean | null;
    sectionStatusList?: Array<ReplySubmissionGroups> | null;
    chpTotalPowerCapacity?: number | null;
    chpMaxHeat?: number | null;
    totalPowCapUnderMaxHeatConds?: number | null;
    totalKweNormRun?: number | null;
    totalKwthNormRun?: number | null;
    schemeLineDiagramComments?: string | null;
    energyFlowDiagramComments?: string | null;
    annualHeatProfileComments?: string | null;
    dailyHeatProfileComments?: string | null;
    heatLoadDurationCurveComments?: string | null;
    rpComments?: string | null;
    hoursOfOperation?: number | null;
    months?: number | null;
    totalFuelEnergyInputs?: number | null;
    estimatedTotalFuelEnergyPrimeEngines?: number | null;
    estimatedTotalFuelEnergyBoilers?: number | null;
    totalPowerGenerated?: number | null;
    totalPowerExported?: number | null;
    totalPowerImported?: number | null;
    heatRejectionFacility?: boolean | null;
    qualifyingHeatOutput?: number | null;
    totalHeatExported?: number | null;
    estimatedTotalHeatOutputUsedInthePrimeMovers?: number | null;
    estimatedTotalHeatOutputUsedIntheBoilers?: number | null;
    powerEfficiencyThreshold?: number | null;
    achivePowerEfficiencyThreshold?: boolean | null;
    powerEfficiency?: number | null;
    heatEfficiency?: number | null;
    sosCertificate?: boolean | null;
    rocsCertificate?: boolean | null;
    cfdCertificate?: boolean | null;
    sumFnX?: number | null;
    sumFnY?: number | null;
    qualityIndex?: number | null;
    qualityIndexThreshold?: number | null;
    qualityIndexThresholdAchived?: boolean | null;
    annualClimateChangeLevyAmount?: number | null;
    annualCarbonPriceSupportAmount?: number | null;
    annualRenewableHeatIncentiveUpliftAmount?: number | null;
    annualRenewablesObligationCertificateAmount?: number | null;
    annualContractsForDifferenceAmount?: number | null;
    annualBusinessRatesReductionAmount?: number | null;
    foi?: number | null;
    fop?: number | null;
    foh?: number | null;
    uncertaintyFactorComment?: string | null;
    rocscfdSumFnX?: number | null;
    rocscfdsumFnY?: number | null;
    rocscfdQualityIndex?: number | null;
    rocscfdQualityIndexThresholdAchived?: boolean | null;
    zRatioDetermined?: boolean | null;
    possibleToDetermineZRatio?: string | null;
    steamExportPressure?: number | null;
    steamturbinesize?: number | null;
    zratio?: number | null;
}